addMueble method in MenuMuebles.java

+++ mueble/Mesa.java
@@ -12,7 +12,7 @@ public abstract class Mesa extends Mueble{
     private final int id;
     private static int count = 0;
 
-    protected Mesa(String articleName, int price, Material material) {
+    protected Mesa(String articleName, float price, Material material) {
         super(articleName, price, material);
         id = count;
         count++;
